# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AeroParts Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /seller/analytics:
    get:
      operationId: getSellerAnalytics
      tags: [seller]
      summary: Get detailed analytics (Pro/Enterprise only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerAnalytics"
        "402":
          description: Upgrade required — Pro or Enterprise plan needed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitReachedError"
        "401":
          description: Unauthorized
...
  /rfqs:
    get:
      operationId: getRfqs
      tags: [rfqs]
      summary: List open RFQs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed]
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfqsPage"
    post:
      operationId: createRfq
      tags: [rfqs]
      summary: Post a new RFQ (no auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfqInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
  /rfqs/{id}:
    get:
      operationId: getRfq
      tags: [rfqs]
      summary: Get a single RFQ with its responses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfqDetail"
        "404":
          description: Not found
  /rfqs/{id}/close:
    post:
      operationId: closeRfq
      tags: [rfqs]
      summary: Close an RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
  /rfqs/{id}/responses:
    post:
      operationId: createRfqResponse
      tags: [rfqs]
      summary: Seller responds to an RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfqResponseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfqResponseItem"
        "401":
          description: Unauthorized
  /seller/rfqs:
    get:
      operationId: getSellerRfqStats
      tags: [seller]
      summary: Get RFQ summary stats for seller dashboard
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerRfqStats"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    LimitReachedError:
      type: object
      required: [error, plan, activeListings, listingLimit]
      properties:
        error:
          type: string
        plan:
          type: string
        activeListings:
          type: integer
        listingLimit:
          type: integer
...
    Rfq:
      type: object
      required: [id, buyerName, partNumber, description, quantity, status, createdAt, accessLevel]
      properties:
        id:
          type: integer
        buyerName:
          type: string
        buyerEmail:
          type: ["string", "null"]
          description: Masked to null for free-plan sellers; full value for Pro/Enterprise
        buyerCompany:
          type: ["string", "null"]
          description: Masked to null for free-plan sellers
        buyerPhone:
          type: ["string", "null"]
          description: Masked to null for free-plan sellers
        partNumber:
          type: string
        description:
          type: string
        aircraftApplicability:
          type: ["string", "null"]
        condition:
          type: ["string", "null"]
        quantity:
          type: integer
        status:
          type: string
          enum: [open, closed]
        accessLevel:
          type: string
          enum: [full, limited]
          description: "full = Pro/Enterprise (complete buyer contact); limited = Free (contact info hidden)"
        createdAt:
          type: string
        updatedAt:
          type: string
    RfqResponseItem:
      type: object
      required: [id, rfqId, sellerId, message, createdAt]
      properties:
        id:
          type: integer
        rfqId:
          type: integer
        sellerId:
          type: integer
        sellerCompanyName:
          type: string
        message:
          type: string
        listingId:
          type: ["integer", "null"]
        listingPartNumber:
          type: ["string", "null"]
        createdAt:
          type: string
    RfqDetail:
      type: object
      required: [rfq, responses]
      properties:
        rfq:
          $ref: "#/components/schemas/Rfq"
        responses:
          type: array
          items:
            $ref: "#/components/schemas/RfqResponseItem"
    RfqInput:
      type: object
      required: [buyerName, buyerEmail, partNumber, description, quantity]
      properties:
        buyerName:
          type: string
        buyerEmail:
          type: string
        buyerCompany:
          type: ["string", "null"]
        buyerPhone:
          type: ["string", "null"]
        partNumber:
          type: string
        description:
          type: string
        aircraftApplicability:
          type: ["string", "null"]
        condition:
          type: ["string", "null"]
        quantity:
          type: integer
          minimum: 1
    RfqResponseInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        listingId:
          type: ["integer", "null"]
    RfqsPage:
      type: object
      required: [rfqs, total, page, limit]
      properties:
        rfqs:
          type: array
          items:
            $ref: "#/components/schemas/Rfq"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SellerRfqStats:
      type: object
      required: [openRfqs, myResponses]
      properties:
        openRfqs:
          type: integer
        myResponses:
          type: integer
    SellerAnalytics:
      type: object
      required: [plan, inquiryTrend, topListings, rfqResponses, rfqResponseRate]
      properties:
        plan:
          type: string
          enum: [free, pro, enterprise, mro_verified, mro_premium]
        inquiryTrend:
          type: array
          description: Daily inquiry counts for the last 30 days
          items:
            type: object
            required: [date, count]
            properties:
              date:
                type: string
              count:
                type: integer
        topListings:
          type: array
          description: Top 5 listings by inquiry volume
          items:
            type: object
            required: [listingId, partNumber, inquiries]
            properties:
              listingId:
                type: integer
              partNumber:
                type: string
              description:
                type: string
              inquiries:
                type: integer
        rfqResponses:
          type: integer
          description: Total RFQ responses submitted (Enterprise only — 0 for Pro)
        rfqResponseRate:
          type: number
          description: Ratio of responded RFQs to total open RFQs (Enterprise only — 0 for Pro)
...